+++ components/sensors/include/sensors.h
@@ -3,6 +3,4 @@
 #pragma once
 
 _Noreturn void dht22_task(void *pvParameters);
-
-_Noreturn
-void sds011_task(void *pvParameters);
+_Noreturn void sds011_task(void *pvParameters);