+++ CHANGELOG.md
@@ -4,8 +4,9 @@ This changelog suppose to follow rules defined in the [changelog.md](https://cha
 
 ## 0.1.0 - 2024-06-10
 
-Initial release with support for DHT22 and SDS011.
+Initial release with support for DHT22 and SDS011 with primitive synchronization mechanism using WiFi and MQTT.
 
+- **Added**: WiFi / MQTT sync 
 - **Added**: DHT22 support
 - **Added**: SDS011 support (query mode with sleep states)
 - **Added**: Shared memory section with mutex access